Revert "Initial implementation of state shim"
This reverts commit a5ad5ccd4324ee8b4c0ad59f51ed0b0e07846a39.
