+++ cmd/state/internal/cmdtree/cmdtree.go
@@ -107,7 +107,6 @@ func New(prime *primer.Values) *CmdTree {
 		newInviteCommand(prime),
 		tutorialCmd,
 		newPrepareCommand(prime),
-		newShimCommand(prime),
 	)
 
 	applyLegacyChildren(stateCmd, globals)
